# apps/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "RAGBot API",
    "description": "Workspace-scoped REST API for the RAGBot platform. Multi-tenant via workspace_id isolation. Auth uses httpOnly cookies (access_token, refresh_token). RBAC enforced via 31 permissions across 8 resources.\n\nEmail verification is required before signin. Account lockout applies after 5 consecutive failed signin attempts (15 minutes).\n\nTwo-factor auth (TOTP authenticator + one-time backup codes + email-OTP fallback) is per-user opt-in and gates sign-in: when enabled, `/signin` returns `{ mfa_required: true }` and a short-lived (5-min) `2fa_challenge` cookie instead of tokens, and sign-in is completed via `/signin/2fa`.\n\n**Auth convention**: Protected routes read `access_token` cookie. Workspace-scoped routes also require valid workspace membership; the permission required is noted in each endpoint description.",
    "version": "3.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development"
    }
  ],
...
  "security": [
    {
      "AccessToken": []
    }
  ],
  "paths": {
...
    "/api/auth/2fa/backup-codes/regenerate": {
      "post": {
        "tags": ["Two-Factor"],
        "summary": "Regenerate backup codes",
        "description": "Replaces all existing backup codes with a fresh set of 10, returned **once**, invalidating the previous set. Requires re-authentication with both the account password and a current second factor (TOTP or backup code).",
        "operationId": "regenerateBackupCodes",
        "security": [{ "AccessToken": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordCodeReauthRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "New backup codes returned once",
            "headers": {
              "X-Request-Id": {
                "$ref": "#/components/headers/X-Request-Id"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/BackupCodes"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "AccessToken": {
        "type": "apiKey",
        "in": "cookie",
        "name": "access_token",
        "description": "JWT access token (httpOnly cookie set by server, expires in 15m)"
      },
      "RefreshToken": {
        "type": "apiKey",
        "in": "cookie",
        "name": "refresh_token",
        "description": "JWT refresh token (httpOnly cookie set by server, expires in 7d). Used for token rotation — revokes old token and issues new pair."
      },
      "ChallengeToken": {
        "type": "apiKey",
        "in": "cookie",
        "name": "2fa_challenge",
        "description": "Short-lived (5-min) JWT (type `2fa_challenge`, httpOnly cookie set by /signin when 2FA is enabled). Authorizes only the sign-in 2FA completion endpoints; cannot be used as an access token."
      }
    },
    "headers": {
      "X-Request-Id": {
        "description": "Correlation ID for request tracing. Accepts a valid UUID or generates one.",
        "schema": {
          "type": "string",
          "format": "uuid"
        }
      }
    },
...
    "responses": {
      "BadRequest": {
        "description": "Invalid request data",
        "headers": {
          "X-Request-Id": {
            "$ref": "#/components/headers/X-Request-Id"
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "message": "\"email\" is required",
              "data": null
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Authentication required or invalid credentials",
        "headers": {
          "X-Request-Id": {
            "$ref": "#/components/headers/X-Request-Id"
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "examples": {
              "noToken": {
                "value": {
                  "message": "No token provided",
                  "data": null
                }
              },
              "expired": {
                "value": {
                  "message": "Token expired",
                  "data": null
                }
              },
              "invalid": {
                "value": {
                  "message": "Invalid token",
                  "data": null
                }
              }
            }
          }
        }
      },
...
      "NotFound": {
        "description": "Resource not found",
        "headers": {
          "X-Request-Id": {
            "$ref": "#/components/headers/X-Request-Id"
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "message": "Resource not found",
              "data": null
            }
          }
        }
      },
...
    },
    "schemas": {
      "ApiResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "data": {
            "nullable": true
          }
        },
        "required": ["message"]
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "data": {
            "type": "object",
            "nullable": true,
            "example": null
          }
        },
        "required": ["message"]
      },
...
      "PasswordCodeReauthRequest": {
        "type": "object",
        "description": "Re-authentication with the account password plus a current second factor (TOTP or one-time backup code)",
        "properties": {
          "password": {
            "type": "string"
          },
          "code": {
            "type": "string",
            "description": "A current TOTP code or one-time backup code"
          }
        },
        "required": ["password", "code"]
      },
...
      "BackupCodes": {
        "type": "object",
        "description": "One-time backup codes — returned only once, on activate/regenerate",
        "properties": {
          "backup_codes": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "10 single-use backup codes"
          }
        },
        "required": ["backup_codes"]
      },
...
    }
  }
}